Create CBSCWD 1(30 Years of AR Rahman).html
Week 1: Introduction to the Internet and Web Development

Assigned task:- 
Build a fun webpage with HTML that includes not only text and images but also an embedded YouTube video or a Spotify playlist related to your chosen topic. Make it engaging and enjoyable to read and explore. 

I have created a webpage celebrating the 30 years of Rahmanian Music for this assignment